Bug 2013273 - Make nsMencloseNotation an enum class. r=layout-reviewers,dshin
Differential Revision: https://phabricator.services.mozilla.com/D280936

UltraBlame original commit: a6685da44ec6a3eefd9cbbde9e9f724e46db84b6

+++ layout/mathml/nsMathMLFrame.h
@@ -274,7 +274,7 @@ class nsMathMLFrame : public nsIMathMLFrame {
 
   void DisplayBar(mozilla::nsDisplayListBuilder* aBuilder, nsIFrame* aFrame,
                   const nsRect& aRect, const mozilla::nsDisplayListSet& aLists,
-                  uint32_t aIndex = 0);
+                  uint16_t aIndex = 0);
 
   
   nsPresentationData mPresentationData;
